Merge flags `httpserveraddress` and `mode` into one (`http`).
The program runs in CLI mode by default. If `--http` flag is set, it
should switch to HTTP mode. Using `-http` to specify the address is
inspired by godoc's --http flag.
This change requires changes to the config file format: the `mode`
setting shall be removed, and the `httpserveraddress` setting shall be
renamed to `http`.