detect if dcrcli.conf is not configured #50
If the config file is not found and the user did not pass Dcrwallet credentials in command line options, propose to guide the user in creating a new config file and do same if the user accepts, else quit the application
use text/template for the conf parsing


Clean up


Default value for RPC credentails

Suggest and use default value for the RPC credentials
Refactored the code to use the already existing RequestInput

Modified the terminalpromt.RequestInput func to eccept nil validation func
Cleaning up


remove empty validator

define default values as constants

configure handler

moved the config setup to a handler and inform the user of the appropriate configuration command at startup if config not found. removed the sample-dcrcli.conf to leave the template string in the handler as the single source of truth for the config file
Requested changes

Renamed 'configure' command to 'config'. removed appName from cli struct and use AppName fn in its place. detached invalidCommandReceived from cli struct. removed unnecessary create config dialog. point out the default values. 
Created a config flag and call the call configure 

Moved the configure func to config.go and also call the func on start if the config file is not found
Created a config flag and call the call configure 

Moved the configure func to config.go and also call the func on start if the config file is not found
Merge remote-tracking branch 'origin/iss-50-conf-detection' into iss-50-conf-detection
configuration instruction in readMe