requested changes
Resolved the issue of 'Total change amount percentage must be equal to 100. Current total is 101'.
Corrected the logic the checking unconfirmed balance.
moved error notification close to the 'Next Button' on the send page.
Re-ordered the input selection functions and updated the send page to not show change customization if sending max amount