Replace missing dependency with a compatible version.
Missing dependencies between godcr and dcrlibwallet caused build failures.
I found that the commit history on dcrlibwallet is inconsistent,
most likely owing to a forced push (using git push --force) that overwrote the history.
I found a branch with similar state as what was expected by godcr at the state of the PR I was working on.
This gave me a clue to where I can find a compatible version of dcrlibwallet as the lost versions.

What commit clobbered the history and deleted the reference `90259892e503` is inconclusive, but I've got the build working now.